begin process of eliminating unneccesary enigma class
