+++ lib/enigma.rb
@@ -2,29 +2,37 @@ require_relative 'enigma_helper'
 require 'date'
 
 class Enigma
-  include Cryptable
 
-  def encrypt(message, key = keygen, date = dategen)
-    final_offsets = generate_final_offsets(key, date)
-    message = message.downcase.split('')
-    output = message.map do |character|
-      build_encrypt_array(character, output, final_offsets)
-    end
-    encrypt_out(output, key, date)
-  end
+  attr_reader :today_date,
+              :character_set
 
-  def decrypt(message, key, date = dategen)
-    final_offsets = generate_final_offsets(key, date)
-    message = message.downcase.split('')
-    output = message.map do |character|
-      build_decrypt_array(character, output, final_offsets)
-    end
-    decrypt_out(output, key, date)
-  end
+  # def initialize
+  #   @today_date = Date.today.strftime('%d%m%y')
+  #   @character_set = ('a'..'z').to_a << ' '
+  # end
 
-  def crack(message, date = dategen)
-    message.downcase!
-    key = decipher_key(message, date)
-    decrypt(message, key, date)
-  end
+  # def keygen
+  #   (Array.new(5) { rand(0..9) }).join
+  # end
+
+  # def key_array(key)
+  #   [
+  #     key[0..1].to_i,
+  #     key[1..2].to_i,
+  #     key[2..3].to_i,
+  #     key[3..4].to_i
+  #   ]
+  # end
+
+  # def offset_array(date)
+  #   sq_date = date.to_i**2
+  #   sq_date_ary = sq_date.to_s[-4..].split("")
+  #   sq_date_ary.map { |num| num.to_i }
+  # end
+
+  # def generate_final_offsets(key, date)
+  #   keys = key_array(key)
+  #   offsets = offset_array(date)
+  #   keys.zip(offsets).map(&:sum)
+  # end
 end
